Extend init-runonce.sh script - create addiditonal resources
For the upstream CI, we want to use Rally and specifically
Tempest tests. However, we want to run Tempest through static
accounts instead of dynamic ones, as it is unnecessary and
extends the testing time. Ideally, we also do not want to use
the admin project and admin user, since it has an admin role
and some tests might not function correctly because of this.

Therefore, it is necessary to add a project, user, and all its
resources so that we can run the tests. I achieved this by
modifying the init-runonce.sh script, which now behaves exactly
as before but has new features if the user calls with parameters.

Outside of CI, this is also useful for the initialization of
resources outside the admin project during a fresh installation
of OpenStack.

Change-Id: I5efbb7bc49da0eaba0eba2e909651dc91b96191e